+++ src/modules/business/services/businessContact/DeleteBusinessContactService.ts
@@ -6,6 +6,7 @@ import IBusinessContactRepository from '../../repositories/IBusinessContactRepos
 
 interface IRequest {
   business_contact_id: string;
+  business_id: string;
 }
 
 @injectable()
@@ -15,13 +16,23 @@ export default class DeleteBusinessContactService {
     private businessContactRepository: IBusinessContactRepository,
   ) {}
 
-  public async execute({ business_contact_id }: IRequest): Promise<void> {
+  public async execute({
+    business_id,
+    business_contact_id,
+  }: IRequest): Promise<void> {
     const businessContact = await this.businessContactRepository.findById(
       business_contact_id,
     );
 
     if (!businessContact) {
       throw new AppError('Business not found.', 404);
+    }
+
+    if (businessContact.business_id !== business_id) {
+      throw new AppError(
+        'Business does not have permission to delete this contact.',
+        403,
+      );
     }
 
     await this.businessContactRepository.delete(business_contact_id);
