+++ src/modules/business/services/businessLocation/DeleteBusinessLocationService.ts
@@ -6,6 +6,7 @@ import IBusinessLocationRepository from '../../repositories/IBusinessLocationRep
 
 interface IRequest {
   business_location_id: string;
+  business_id: string;
 }
 
 @injectable()
@@ -15,13 +16,23 @@ export default class DeleteBusinessLocationService {
     private businessLocationRepository: IBusinessLocationRepository,
   ) {}
 
-  public async execute({ business_location_id }: IRequest): Promise<void> {
+  public async execute({
+    business_id,
+    business_location_id,
+  }: IRequest): Promise<void> {
     const businessLocation = await this.businessLocationRepository.findById(
       business_location_id,
     );
 
     if (!businessLocation) {
       throw new AppError('Business not found.', 404);
+    }
+
+    if (businessLocation.business_id !== business_id) {
+      throw new AppError(
+        'Business does not have permission to delete this location.',
+        403,
+      );
     }
 
     await this.businessLocationRepository.delete(business_location_id);
