+++ src/modules/business/services/businessProduct/DeleteBusinessProductService.ts
@@ -6,6 +6,7 @@ import IBusinessProductRepository from '../../repositories/IBusinessProductRepos
 
 interface IRequest {
   business_product_id: string;
+  business_id: string;
 }
 
 @injectable()
@@ -15,13 +16,23 @@ export default class DeleteBusinessProductService {
     private businessProductRepository: IBusinessProductRepository,
   ) {}
 
-  public async execute({ business_product_id }: IRequest): Promise<void> {
+  public async execute({
+    business_id,
+    business_product_id,
+  }: IRequest): Promise<void> {
     const businessProduct = await this.businessProductRepository.findById(
       business_product_id,
     );
 
     if (!businessProduct) {
       throw new AppError('Business not found.', 404);
+    }
+
+    if (businessProduct.business_id !== business_id) {
+      throw new AppError(
+        'Business does not have permission to delete this product.',
+        403,
+      );
     }
 
     await this.businessProductRepository.delete(business_product_id);
