+++ src/modules/business/services/eventBusiness/DeleteEventBusinessService.ts
@@ -6,6 +6,7 @@ import IEventBusinessRepository from '../../repositories/IEventBusinessRepositor
 
 interface IRequest {
   event_business_id: string;
+  business_id: string;
 }
 
 @injectable()
@@ -15,13 +16,23 @@ export default class DeleteEventBusinessService {
     private eventBusinessRepository: IEventBusinessRepository,
   ) {}
 
-  public async execute({ event_business_id }: IRequest): Promise<void> {
+  public async execute({
+    business_id,
+    event_business_id,
+  }: IRequest): Promise<void> {
     const eventBusiness = await this.eventBusinessRepository.findById(
       event_business_id,
     );
 
     if (!eventBusiness) {
       throw new AppError('Event not found.', 404);
+    }
+
+    if (eventBusiness.business_id !== business_id) {
+      throw new AppError(
+        'Business does not have permission to delete this event.',
+        403,
+      );
     }
 
     await this.eventBusinessRepository.delete(event_business_id);
